# backend/src/main/java/io/methodra/backend/api/AuthController.java
package io.methodra.backend.api;

import io.methodra.backend.auth.AuthService;
import jakarta.validation.Valid;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/v1/auth")
public class AuthController {
    private final AuthService auth;

    public AuthController(AuthService auth) { this.auth = auth; }

    @PostMapping("/register")
    @ResponseStatus(HttpStatus.CREATED)
    public AuthService.AuthResult register(@Valid @RequestBody Credentials request) {
        return auth.register(request.email(), request.password());
    }

    @PostMapping("/login")
    public AuthService.AuthResult login(@Valid @RequestBody Credentials request) {
        return auth.login(request.email(), request.password());
    }

    public record Credentials(
            @NotBlank @Email @Size(max = 320) String email,
            @NotBlank @Size(min = 10, max = 128) String password
    ) {}
}