# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Methodra API
  version: 1.0.0-demo
  description: Optional Methodra catalog, authentication and conflict-safe sync backend.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/auth/register:
    post:
      summary: Create an optional Methodra account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '201':
          description: Account and bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
  /api/v1/auth/login:
    post:
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    Credentials:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email, maxLength: 320}
        password: {type: string, minLength: 10, maxLength: 128}
    AuthResult:
      type: object
      properties:
        userId: {type: string, format: uuid}
        email: {type: string, format: email}
        token: {type: string}
        expiresAt: {type: string, format: date-time}
...